refactor: changed action name to "sendMobilePush"
